Include the 682f6e0 bug in the unit tests.
The bug was that we considered "\nr" as the beginning of a status entry.
However, there are crypto lines starting with "r", so we need to look for
"\nr ".  Now there are crypto lines starting with "r" in the unit tests.
Doesn't fix the problem, but helps discover it in the future.  Fix comes
next.