Fix download logic a bit.
Now we report a failed request instead of silently ignoring it.  As a
result, we don't wait forever if we cannot download a single consensus.

The download logic is still somewhat broken.  We don't wait for all
consensus requests to complete or fail, which results in adding (failed)
requests to the queue when we think we're done.  This is probably
harmless, but should be fixed at some point.