Fix a problem with downloading relay descriptors.
The consensus-health checker recently reported quite often that it
couldn't download any consensuses.  What happened was that the first
directory authority immediately with an error.  We decided that there's
nothing to download anymore, because a) we requested the consensus from
all authorities and didn't know about any votes yet that we could
download.  However, what we need to do is wait until all running consensus
downloads finish.