Add a config option to handle unrecognized lines.
Most applications don't care about unrecognized lines in descriptors as
long as the lines they care about are contained.  But some applications
rather want to skip a descriptor with unrecognized lines, e.g., the bridge
descriptor sanitizer.

Add a switch to downloaders and readers to control the desired behavior,
that is, to either fail parsing or ignore unrecognized lines.  In the
latter case, unrecognized are written to a list for later inspection by
the application.

Corrupt lines with known keywords always lead to failing the parse step.