Remove two suggested data sources from the TODO list.
Using Tor's control port as a data source could be useful in some cases,
but Damian and I decided it's not worth the effort.  We wouldn't learn
much new stuff compared to running a Tor client/relay and reading the
cached-* files.  Damian says that a good workaround for file permission
issues with cached-* files is to run the Tor process and Java process as
the same user.

Having a database as a data source was a nice idea, too, but we don't have
such a database.  Once we have a good database we can still write a
metrics-lib wrapper for it.