Make parse history in descriptor reader more accessible.
So far, the only way to prevent files from being parsed repeatedly in
distinct runs was to specify a history file that only metrics-lib was
supposed to read and write.  However, some applications may want to
specify the list of files to exclude themselves, or they may want to learn
which files have been excluded and which have been parsed.  These
applications shouldn't be forced to mess with the history file.

Add three methods to the descriptor reader for these applications.  They
should also play nicely together with the history file approach.

AFAIK, stem has methods with the same purpose but a slightly different
semantic.