Fix build errors on Debian systems.
The local lib directory is not used anymore and respective references
were removed. The java dependencies are now specified in the build.xml
and taken from their installed locations.

In addition to git, openjdk-6-jdk and ant the following java packages
have to be installed:
 - libcommons-codec-java
 - libcommons-compress-java
 - junit4

Minor tweaks by Karsten Loesing.