Fix encoding problem when parsing multi-descriptor files.
When we're parsing a descriptor file with potentially more than one
descriptor in it, we're converting file contents to String to be able to
search for descriptor beginnings using String methods.  But we're not
passing a character encoding, leaving it up to Java to guess.  What we
should do is tell it to use "US-ASCII" as encoding, which is sufficient to
find keywords marking the beginning of a new descriptor.

Fixes #11821.