Store bandwidth histories more efficiently.
We were storing bandwidth histories in TreeMap<Long, Long>() with keys
being time in millis and values being bandwidth values.  This showed up in
profiles.  It's far more (memory-)efficient to store bandwidth values in a
long[] and put together the TreeMap when the caller requests it.  And if
the bandwidth history is evaluated exactly once, there should not even be
a CPU overhead.