Avoid parsing descriptor contents to Maps.
Extra-info descriptors contain lots of comma-separated key=value lists
that we store in SortedMap instances.  But those occupy a lot of memory,
and it's not certain that we'll ever want to use the contained keys or
values.

New approach: when parsing a descriptor, use regular expressions to check
if lines are valid, and delay parsing into maps until needed.