Replace JAXB with Apache Commons Codec dependency.
The reason is that JAXB won't be part of standard Java 9 anymore, and
the only reason why we used it was to get rid of Apache Commons Codec
as dependency, three years ago. Now it makes sense to just go back to
Apache Commons Codec in preparation of upgrading to Java 9.

Partially reverts 1734d16.

Fixes #25444.