Resume previously aborted downloads.
So far, whenever DescriptorCollector fetches a remote file, it writes
its contents to a local temporary file and later renames that file.
The idea is to avoid incomplete downloads. However, if a local
temporary file already exists from an earlier run, DescriptorCollector
fails and skips the remote file. This is not intended.

With this change, DescriptorCollector always overwrites a local
temporary file when it finds one. It still doesn't delete incomplete
temporary files that are not re-attempted later. But not skipping them
is already a good start.

Fixes #24153.